+++ src/pages/index.tsx
@@ -1,4 +1,4 @@
-import { Button } from '@/components/Button/Button';
+import { Button } from '@/components/inputs/Button/Button';
 
 function HomePage() {
   return (
