+++ src/presentation/components/Snackbar/SnackbarProvider.stories.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { createStoryMeta } from 'src/utils/storybookUtils';
-import { TextButton } from '../TextButton/TextButton';
+import { TextButton } from '../inputs/TextButton/TextButton';
 import { SnackbarProvider, useSnackBar } from './SnackbarProvider';
 
 export default createStoryMeta(SnackbarProvider, {
