+++ src/presentation/components/Tooltip/Tooltip.stories.tsx
@@ -1,7 +1,7 @@
 import { useToggle } from 'src/hooks/useToggle';
 import { ComponentStoryObj, createStoryMeta } from 'src/utils/storybookUtils';
 import styled from 'styled-components';
-import { TextButton } from '../TextButton/TextButton';
+import { TextButton } from '../inputs/TextButton/TextButton';
 import { Tooltip, TooltipProps } from './Tooltip';
 
 const Wrapper = styled.div`
