+++ src/presentation/components/dialogs/DialogProvider.stories.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { createStoryMeta } from 'src/utils/storybookUtils';
-import { TextButton } from '../TextButton/TextButton';
+import { TextButton } from '../inputs/TextButton/TextButton';
 import { CustomDialog } from './CustomDialog/CustomDialog';
 import { DialogProvider, useDialog } from './DialogProvider';
