+++ src/presentation/components/inputs/Button/Button.test.tsx
@@ -1,4 +1,4 @@
-import { Button } from '@/components/Button/Button';
+import { Button } from '@/components/inputs/Button/Button';
 import { renderWithStyle } from 'src/utils/testUtil';
 
 describe('Button', () => {
